Move the deals into the Deals directory
