+++ src/Teamleader/Client.php
@@ -8,9 +8,9 @@ use Teamleader\Entities\CRM\Company;
 use Teamleader\Entities\CRM\Contact;
 use Teamleader\Entities\CRM\BusinessType;
 use Teamleader\Entities\CRM\Tag;
-use Teamleader\Entities\Deal;
-use Teamleader\Entities\DealPhase;
-use Teamleader\Entities\DealSource;
+use Teamleader\Entities\Deals\Deal;
+use Teamleader\Entities\Deals\DealPhase;
+use Teamleader\Entities\Deals\DealSource;
 use Teamleader\Entities\General\Department;
 use Teamleader\Entities\General\User;
 use Teamleader\Entities\Invoicing\CreditNote;
